+++ lib/bal/model.js
@@ -1,7 +1,5 @@
-/* eslint camelcase: off */
-import {pick, find, filter, max} from 'lodash'
-import {unparse} from 'papaparse'
-import proj from '@etalab/project-legal'
+import {find, max} from 'lodash'
+import {exportAsCsv} from './export-csv'
 
 function resetItem(item) {
   item.modified = null
@@ -11,124 +9,6 @@ function resetItem(item) {
 
 function findVoieNameInCommune(commune, name) {
   return find(commune.voies, voie => voie.nomVoie === name)
-}
-
-function filterActiveItems(items) {
-  return filter(items, item => !item.deleted)
-}
-
-const COMMUNE_PROPERTIES = ['code', 'nom']
-const VOIE_PROPERTIES = ['idVoie', 'codeVoie', 'nomVoie', 'source', 'dateMAJ']
-const NUMERO_PROPERTIES = ['id', 'numeroComplet', 'numero', 'suffixe', 'source', 'dateMAJ']
-const POSITION_PROPERTIES = ['type', 'coords', 'dateMAJ', 'source']
-
-function computeProperties(item, keys) {
-  const initialProperties = pick(item, keys)
-  return item.modified ? {...initialProperties, ...item.modified} : initialProperties
-}
-
-function applyChanges(tree) {
-  const communes = filterActiveItems(tree.communes || [])
-    .map(commune => {
-      const voies = filterActiveItems(commune.voies || [])
-        .map(voie => {
-          const numeros = filterActiveItems(voie.numeros || [])
-            .map(numero => {
-              const positions = filterActiveItems(numero.positions || [])
-                .map(position => computeProperties(position, POSITION_PROPERTIES))
-              return {...computeProperties(numero, NUMERO_PROPERTIES), positions}
-            })
-          const position = voie.position ? computeProperties(voie.position, POSITION_PROPERTIES) : {}
-          return {...computeProperties(voie, VOIE_PROPERTIES), numeros, position}
-        })
-      return {...computeProperties(commune, COMMUNE_PROPERTIES), voies}
-    })
-  return {communes}
-}
-
-function splitNumeroComplet(numeroComplet) {
-  if (!numeroComplet) {
-    return {numero: '', suffixe: ''}
-  }
-  const result = numeroComplet.match(/^(\d+)/g)
-  if (!result) {
-    throw new Error('Numero complet non valide : ' + numeroComplet)
-  }
-  const [numero] = result
-  const suffixe = numero.length < numeroComplet.length ? numeroComplet.substr(numero.length) : ''
-  return {numero, suffixe}
-}
-
-function roundCoordinate(coordinate, precision) {
-  return parseFloat(coordinate.toFixed(precision))
-}
-
-function createRow(obj) {
-  const row = {
-    cle_interop: '', // TODO
-    uid_adress: '',
-    voie_nom: obj.nomVoie,
-    ...splitNumeroComplet(obj.numeroComplet),
-    commune_nom: obj.nomCommune,
-    position: '',
-    long: '',
-    lat: '',
-    x: '',
-    y: '',
-    source: '', // TODO
-    date_der_maj: obj.dateMAJ // TODO
-  }
-  if (obj.position && obj.position.coords) {
-    const {coords, type} = obj.position
-    const projectedCoords = proj(coords)
-    row.position = type
-    row.long = roundCoordinate(coords[0], 6)
-    row.lat = roundCoordinate(coords[1], 6)
-    if (projectedCoords) {
-      row.x = projectedCoords[0]
-      row.y = projectedCoords[1]
-    }
-  }
-  return row
-}
-
-function serializeCompactTree(tree) {
-  const rows = []
-  tree.communes.forEach(commune => {
-    const communeCtx = {codeCommune: commune.code, nomCommune: commune.nom}
-    commune.voies.forEach(voie => {
-      const voieCtx = {
-        ...communeCtx,
-        idVoie: voie.idVoie,
-        nomVoie: voie.nomVoie,
-        codeVoie: voie.codeVoie,
-        source: voie.source, // TODO: Merge
-        dateMAJ: voie.dateMAJ // TODO: Merge
-      }
-      if (voie.numeros.length > 0) {
-        voie.numeros.forEach(numero => {
-          const numeroCtx = {
-            ...voieCtx,
-            numeroComplet: numero.numeroComplet,
-            source: voie.source, // TODO: Merge
-            dateMAJ: voie.dateMAJ // TODO: Merge
-          }
-          if (numero.positions.length > 0) {
-            numero.positions.forEach(position => {
-              rows.push(createRow({...numeroCtx, position}))
-            })
-          } else {
-            rows.push(createRow(numeroCtx))
-          }
-        })
-      } else if (voie.position) {
-        rows.push(createRow({...voieCtx, numeroComplet: '99999', position: voie.position}))
-      } else {
-        rows.push(createRow({...voieCtx, numeroComplet: '99999'}))
-      }
-    })
-  })
-  return rows
 }
 
 export default class BAL {
@@ -465,11 +345,6 @@ export default class BAL {
   }
 
   async exportAsCsv() {
-    const unparseOptions = {
-      delimiter: ';',
-      header: true,
-      newline: '\r\n'
-    }
-    return unparse(serializeCompactTree(applyChanges({communes: this.communes})), unparseOptions)
+    return exportAsCsv({communes: this.communes})
   }
 }
